default workdir/outdir; workdir cleanup; changes for dashboard
