+++ src/tiledb/cloud/workflows/nextflow/__init__.py
@@ -1,16 +1,20 @@
+from ..common import default_outdir
+from ..common import default_workdir
 from .history import consolidate_history
 from .history import delete_history
+from .history import get_history
 from .history import get_log
-from .history import list_history
 from .register import register
 from .run import resume
 from .run import run
 
 __all__ = [
-    "list_history",
-    "get_log",
+    "default_outdir",
+    "default_workdir",
     "consolidate_history",
     "delete_history",
+    "get_log",
+    "get_history",
     "register",
     "run",
     "resume",
